requirements.txt
Django==4.2.13
djangorestframework==3.15.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
PyJWT==2.7.0
whitenoise==6.7.0
gunicorn==21.2.0